refactor(audit): split the three files this PR pushed over the line limit
CLAUDE.md keeps files under 800 lines. Pass 1 and 2 pushed three past it, so
they get split along seams that were already there.

- types.h (850) held two concerns that never reference each other: the
  surface/decoration fold and the window animation transitions. Split into
  surface_types.h (403) and transition_types.h (475). types.h stays as a 44-line
  umbrella so no consumer's includes change, and a TU that needs only one half can
  now include that half and skip the other's dependencies.
- desktoptransitionmanager.cpp (833) split at "how a desktop becomes a texture"
  vs "how two textures become a frame" — the capture side moves to
  desktoptransitioncapture.cpp. 649 + 217.
- surfacelayers.cpp (818) gives up surfacePresentShader() to decoration_render.cpp,
  which already owns both of the program's consumers. 698 + 617. That also undoes
  an SRP smear: a compile-and-cache routine living in the render TU while its
  siblings live in surface_compile.cpp.

Pure code moves. No logic altered.

Also fixed along the way, from the pass-2 reports:
- The teardown timer now re-checks the transition's own clock instead of trusting
  the delay it was armed with. Restore suppression rebases startTimeMs every
  withheld frame, so a timer fixed at install fired while the open animation still
  had up to 250 ms left, cutting it mid-flight.
- renderSurfaceChainComposite held a QHash iterator across a nested drawWindow
  that can rehash it. Copies the record now — a latent use-after-free.
- bufferShaderPaths was the one uncapped axis of pack metadata: each pass costs a
  canvas-sized texture and a fullscreen draw per decorated window per frame, and
  anything past the fourth is structurally unreadable since the fold binds only
  iChannel0..3. Capped at the parse boundary.

Still over the limit and DESCOPED, with reason: paint_pipeline.cpp (2238),
shader_transitions.cpp (2156), plasmazoneseffect.h (1845), window_lifecycle.cpp
(1455), lifecycle.cpp (1157), autotilehandler/signals.cpp (870). Every one of
these was already over on v3.2 — paint_pipeline at 2112 and shader_transitions at
2052 are ~2.6x the limit and predate this branch. Decomposing them is a real
architectural change, not an audit drive-by, and doing it here would bury the
timing fixes in a 4000-line move.

264/264 pass.
