+++ kwin-effect/CMakeLists.txt
@@ -82,6 +82,7 @@ set(plasmazones_effect_SRCS
     snaphandler.cpp
     snaphandler.h
     desktoptransitionmanager.cpp
+    desktoptransitioncapture.cpp
     desktoptransitionmanager.h
     navigationhandler.cpp
     navigationhandler.h
@@ -121,6 +122,10 @@ set_source_files_properties(snapassistthumbnailcapture.cpp PROPERTIES SKIP_UNITY
 # fullscreen vertex-source constant; keep it out of the Unity blob so those
 # never ODR-collide with surfacelayers.cpp's equivalents.
 set_source_files_properties(desktoptransitionmanager.cpp PROPERTIES SKIP_UNITY_BUILD_INCLUSION ON)
+# desktoptransitioncapture.cpp holds the capture half of the same class, including
+# its own file-local helpers (captureFormatFor, allocateOutputTexture); same
+# reason, same exclusion.
+set_source_files_properties(desktoptransitioncapture.cpp PROPERTIES SKIP_UNITY_BUILD_INCLUSION ON)
 
 # Link required libraries
 target_link_libraries(kwin_effect_plasmazones PRIVATE
